Radis_Monitoring_Tool/rmon/requirements.txt

#整个项目的依赖软件包

Flask==0.12.2
marshmallow==2.13.6
hiredis==0.2.0
redis==2.10.6
Flask-SQLAlchemy==2.2
ipython==6.2.0
pylint==1.7.2
pytest==3.2.2
pytest-flask==0.10.0
flask-shell-ipython==0.3.0
PyJWT==1.5.3
wechatpy==1.5.1
cryptography==2.1.1

